requirements.txt
apache-airflow>=2.8,<3.0
apache-airflow-providers-postgres>=5.10.0
pytest>=8,<9
